+++ package.json
@@ -14,6 +14,7 @@
     "release:version": "lerna version --no-changelog --no-push --no-git-tag-version --no-private",
     "release:build": "pnpm --filter \"./packages/**\" run build",
     "release:publish": "tsx scripts/publish.mjs",
+    "release:publish-canary": "tsx scripts/publish-canary.mjs",
     "pkg-pr-new-packages": "pnpm ls -r --parseable --depth -1 -F \"./packages/**\""
   },
   "pnpm": {
